% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfem.init.prms.R
\name{bfem.init.prms}
\alias{bfem.init.prms}
\title{Frequentist Mstep to initialize bfem.main and bem.sparse}
\usage{
bfem.init.prms(Y, U, tau, model, method)
}
\description{
Frequentist Mstep to initialize bfem.main and bem.sparse
}
