\name{FisherEM-package}
\alias{FisherEM-package}
\alias{FisherEM}
\docType{package}
\title{The FisherEM Algorithm to Simultaneously Cluster and Visualize High-Dimensional Data}
\description{
FisherEM is an efficient algorithm for the unsupervised classification of high-dimensional data. FisherEM 
models and clusters the data in a discriminative and low-dimensional latent subspace. It also provides a low-dimensional representation of the clustered data. A sparse version of Fisher-EM algorithm is also provided.
}
\details{
\tabular{ll}{
Package: \tab FisherEM\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2012-07-09\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}

}
\author{ 
Charles Bouveyron and Camille Brunet

Maintainer: Charles Bouveyron <charles.bouveyron@gmail.com>
}
\references{
Charles Bouveyron, Camille Brunet (2012), "Simultaneous model-based clustering and visualization in the Fisher discriminative subspace.", Statistics and Computing, 22(1), 301-324.

Charles Bouveyron and Camille Brunet (2014), "Discriminative variable selection for clustering with the sparse Fisher-EM algorithm", Computational Statistics, vol. 29(3-4), pp. 489-513.
}


